reorganized file structure and implemented absolute imports
