+++ client/babel.config.js
@@ -11,8 +11,7 @@ module.exports = function(api) {
           moduleName: "@env",
           path: ".env"
         }
-      ]
-    ]
-    // plugins: ['react-native-reanimated/plugin'],
+      ], 
+    ], 
   };
 };
